LayoutTests/media/modern-media-controls/media-controller/media-controller-auto-hide-rewind-with-mouse-enter-expected.txt

Testing the controls bar reappears when the mouse enters the media.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Video started playing, controls bar is visible by default.
PASS controlsBar.classList.contains('faded') is false

Auto-hide timer has elapsed, controls bar should be faded.
PASS controlsBar.classList.contains('faded') is true

Mouse moved over the media, controls bar should be visible.
PASS controlsBar.classList.contains('faded') is false

Auto-hide timer has elapsed, controls bar should be faded.
PASS controlsBar.classList.contains('faded') is true

PASS successfullyParsed is true

TEST COMPLETE
